+++ src/log.cpp
@@ -43,7 +43,7 @@ void LogSystem::move(long long v, bool incomed){
     return;
 }
 
-bool LogSystem::show(int count, AccountSystem &account){
+bool LogSystem::show(long long count, AccountSystem &account){
     if(account.currentPrivilege() < 7)return false;
     if(count == 0){
         std::cout << std::endl;
